+++ app/upgrades/v16/export_test.go
@@ -6,10 +6,7 @@ import (
 	tokenfactorykeeper "github.com/osmosis-labs/osmosis/v25/x/tokenfactory/keeper"
 )
 
-var (
-	AuthorizedQuoteDenoms = authorizedQuoteDenoms
-	AuthorizedUptimes     = authorizedUptimes
-)
+var AuthorizedUptimes = authorizedUptimes
 
 func UpdateTokenFactoryParams(ctx sdk.Context, tokenFactoryKeeper *tokenfactorykeeper.Keeper) {
 	updateTokenFactoryParams(ctx, tokenFactoryKeeper)