+++ app/upgrades/v23/upgrades_test.go
@@ -47,7 +47,6 @@ func (s *UpgradeTestSuite) TestUpgrade() {
 
 	// Update authorized quote denoms
 	concentratedParams := s.App.ConcentratedLiquidityKeeper.GetParams(s.Ctx)
-	concentratedParams.AuthorizedQuoteDenoms = append(concentratedParams.AuthorizedQuoteDenoms, apptesting.USDC)
 	s.App.ConcentratedLiquidityKeeper.SetParams(s.Ctx, concentratedParams)
 
 	// Create two more concentrated pools with positions